pyproject.toml
[tool.poetry]
name = "les_etl_pipeline"
version = "0.1.0"
description = ""
authors = ["francesco.gastaldello <francesco.gastaldello@algoritmica.ai>"]
exclude = ["src/main*.py"]
packages = [
    { include = "src/**/*.py" },
]

[tool.poetry.dependencies]
python = "^3.9"
google-cloud-storage = "^2.6.0"
delta-spark = "2.1.0"
Cerberus = "^1.3.4"
google-cloud-bigquery = "^3.5.0"
unidecode = "^1.3.6"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.24.0"
pandas = "^2.0.3"
pyarrow = "^12.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
